pyproject.toml
[project]
name = "hp-rag"
version = "0.1.0"
description = "RAG pipeline over Harry Potter text using Claude and ChromaDB"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "sentence-transformers>=3.0.0",
    "chromadb>=0.5.0",
    "datasets>=3.0.0",
    "tiktoken>=0.7.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
